requirements.txt
PyGithub>=1.59.0
pyyaml>=6.0.1
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
jinja2>=3.1.2
markdown>=3.5.1
slack-sdk>=3.21.3
aiohttp>=3.9.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6 